Break out logic for Context class
